Fixed an integer overflow bug.

+++ src/main/java/com/yammer/metrics/core/UniformSample.java
@@ -53,9 +53,9 @@ public class UniformSample implements Sample {
 		if (c <= values.length()) {
 			values.set((int) c - 1, value);
 		} else {
-			final int r = (int) (Math.abs(RANDOM.nextLong()) % c);
+			final long r = Math.abs(RANDOM.nextLong()) % c;
 			if (r < values.length()) {
-				values.set(r, value);
+				values.set((int) r, value);
 			}
 		}
 	}
